+++ examples/capture_position_helper.py
@@ -1,5 +1,5 @@
 import os
-from robot_controller import robot
+from src.robot_controller import robot
 
 # Global Constants
 #drive_path = '129.101.98.214' # Bill
