+++ examples/conveyor/proximity_sensor_example.py
@@ -13,10 +13,8 @@
 # - Created by James Lasso on 6/27/2023
 
 # Imports
-import sys
 import time
-import random
-from robot_controller import robot
+from src.robot_controller import robot
 
 # Global Constants
 #drive_path = '129.101.98.214' # Bill
