+++ examples/joint/CRX10_MOUNT_POSITION.py
@@ -14,10 +14,7 @@
 # - Created by James Lasso on 6/27/2023
 
 # Imports
-import sys
-import time
-import random
-from robot_controller import robot
+from src.robot_controller import robot
 
 # Global Constants
 #drive_path = '129.101.98.214' # Bill
