+++ examples/robots/Bill_example.py
@@ -14,10 +14,8 @@
 # - Created by James Lasso on 6/13/2023
 
 # Imports
-import sys
 import time
-import random
-from robot_controller import robot
+from src.robot_controller import robot
 
 # Global Constants
 drive_path = '129.101.98.214' # Bill
@@ -31,7 +29,7 @@ def main():
     crx10 = robot(drive_path)
 
     # Set robot speed
-    crx10.set_speed(200)
+    crx10.set_speed(300)
 
     loops = 1
     while(loops <= 2):
