+++ examples/robots/connectRoboguide.py
@@ -5,13 +5,11 @@
 
 import sys
 sys.path.append('./pycomm3/pycomm3')
-import struct
 import random
-import time
 
-import FANUCethernetipDriver
+from src import FANUCethernetipDriver
 
-drive_path = '129.101.98.244'    # 50iB Larry 
+drive_path = '129.101.98.199'    # CRX10 Bill Roboguide
 
 # read CURPOS from Robot
 print("------------------------")
@@ -48,17 +46,6 @@ myList = CurPosList
 W_PR_1_return = FANUCethernetipDriver.writeCartesianPositionRegister(drive_path, PRNumber, myList)
 
 print("W_PR_1_return =", W_PR_1_return)
-# write R[5] to set Speed in mm/sec
-print("------------------------")
-print(" write R[5] to set arm speed")
-print("------------------------")
-
-RegNum = 5
-Value = 100
-W_R_5_return = FANUCethernetipDriver.writeR_Register(drive_path, RegNum, Value)
-
-print ("W_R_5_return=",W_R_5_return)
-
 
 
 # write R[1] to start Robot
