+++ src/robot_controller.py
@@ -20,9 +20,6 @@
 # Imports
 import sys
 sys.path.append('./pycomm3/pycomm3')
-import struct
-import random
-import time
 import math
 import FANUCethernetipDriver
